pyproject.toml
[project]
name = "local-unpaywall"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary>=2.9.10",
    "tqdm>=4.64.0",
    "requests>=2.28.0",
    "dotenv>=0.9.9",
    "pytest>=8.4.0",
]
